Fix issue: Duplicate PRIMARY key sqlalchemy.exc.IntegrityError: (pymysql.err.IntegrityError) (1062, "Duplicate entry '2023-09-17 20:21:48-41' for key 'PRIMARY'") [SQL: 'INSERT INTO colortemp5 (timestamp, field, value) VALUES (%(timestamp)s, %(field)s, %(value)s)'] [parameters: {'timestamp': '2023-09-17 20:21:47.811581', 'field': 41, 'value': 0}] (Background on this error at: http://sqlalche.me/e/gkpj)
This is because the first developer, Wang,  had a worst desgin.
The resolution of timestamp is ONE second, and he uses time as  PRIMARY key.
When we use MQTT protocol, the response is very fast; therefore, several sernsors' values are came at the same second (only difference in ms.).
And then duplicate PRIMARY key issue happened. Orz...